requirements.txt
flask
jass-kit
logging
numpy
tensorflow